My first created quiz using python
I made a quiz with the objective being to guess the correct state capital of the state listed in the question.  This is the first part of what I eventually would like to do as a project.  I am trying to figure out how to install this quiz as a mobile app.  I would like for this to be a quiz under the Parental Control settings.  I would like to give parents a chance to set this quiz as the only way to unlock their phone after it's been in use for 10 minutes or more.  Rather than just entering in a 6 digit code, I would rather the child have to answer a question relevant to what they are studying in school.